NAR'SIEN BULLSHIT IS DONE

+++ infinity/code/game/gamemodes/heretic/ritual.dm
@@ -20,14 +20,22 @@
 	var/required_cultists = 1
 	var/ritual_radius = 1 //Basically 3x3 with radius 1
 
+	var/noghosts = 0
+
 /datum/ritual/proc/speak_incantation(var/mob/living/user, var/incantation)
 	var/datum/language/L = all_languages[LANGUAGE_CULT]
 	if(incantation && (L in user.languages))
 		user.say(incantation, L)
 
-/datum/ritual/proc/check_cultists(var/obj/effect/ritual_rune, var/mob/user)
+/datum/ritual/proc/check_cultists(var/obj/effect/rune/ritual_rune, var/mob/user)
 	var/cultists = 0
-	for(var/mob/cultist in orange(ritual_rune, ritual_radius))
+	for(var/mob/living/cultist in orange(ritual_rune, ritual_radius))
+
+		if(ishuman(cultist))
+			var/mob/living/carbon/human/H = cultist
+			if(H.species == "Cult" && noghosts)
+				continue
+
 		if(iscultist(cultist))
 			cultists += 1
 
@@ -36,16 +44,18 @@
 
 	return cultists
 
-/datum/ritual/proc/mass_incantation(var/obj/effect/ritual_rune, var/incantation)
+/datum/ritual/proc/mass_incantation(var/obj/effect/rune/ritual_rune, var/incantation)
 	for(var/mob/cultist in orange(ritual_rune, ritual_radius))
 		if(iscultist(cultist))
 			speak_incantation(cultist, incantation)
 
-/datum/ritual/proc/cast(var/obj/effect/ritual_rune, var/mob/user)
+/datum/ritual/proc/cast(var/obj/effect/rune/ritual_rune, var/mob/user)
 	var/has_knife = 0
 	var/has_tome = 0
 	var/has_robes = 0
 	var/cult_ground = 0
+
+	mass_incantation("Da gla'rar nao'reda!")
 
 	if(istype(user.get_active_hand(), /obj/item/weapon/material/knife) || istype(user.get_inactive_hand(), /obj/item/weapon/material/knife))
 		has_knife = 1
@@ -53,7 +63,7 @@
 	if(istype(user.get_active_hand(), /obj/item/weapon/book/tome) || istype(user.get_inactive_hand(), /obj/item/weapon/book/tome))
 		has_tome = 1
 
-	if(istype(user.get_equipped_item(slot_head), /obj/item/clothing/head/culthood) && istype(user.get_equipped_item(slot_wear_suit), /obj/item/clothing/suit/cultrobes) && istype(user.get_equipped_item(slot_shoes), /obj/item/clothing/shoes/cult))
+	if(istype(user.get_equipped_item(slot_head), /obj/item/clothing/head/culthood) && istype(user.get_equipped_item(slot_wear_suit), /obj/item/clothing/suit/storage/hooded/cultrobes) && istype(user.get_equipped_item(slot_shoes), /obj/item/clothing/shoes/cult))  //INF was /obj/item/clothing/suit/cultrobes
 		has_robes = 1
 
 	if(!istype(ritual_rune, rune_type))
@@ -77,12 +87,14 @@
 		to_chat(user, SPAN_WARNING("The ritual is too complex to perform it without a tome!"))
 		return
 
+	mass_incantation("Ka'anahe ra jab'a'rate!")
+
 	if(ritual_flags & NEEDS_ARMOR && !has_robes)
 		to_chat(user, SPAN_WARNING("You can't manage to perform this ritual without a propper set of clothing!"))
 		return
 
 	if(ritual_flags & NEEDS_FLOOR && !cult_ground)
-		to_chat(user, SPAN_WARNING("Your connection to this world is too strong on this floor!"))
+		to_chat(user, SPAN_WARNING("Your connection to this world is too strong on non-defiled floor!"))
 		return
 
 	for(var/obj in requirments)
@@ -91,16 +103,76 @@
 			return
 
 	for(var/obj in requirments)
-		var/req_item = locate(obj) in get_turf(ritual_rune)
-		qdel(req_item)
+		if(requirments[obj] > 1)
+			var/list/listing = locate(obj) in get_turf(ritual_rune)
+			if(listing.len < requirments[obj])
+				to_chat(user, SPAN_WARNING("You need at least [requirments[obj]] [obj] on the rune to perform this ritual!"))
+				return
+
+
+	sleep(5)
+
+	if(!check_cultists(ritual_rune))
+		return
+
+	mass_incantation("Ya za'nere da per'ene!")
+
+	for(var/obj in requirments)
+		if(requirments[obj])
+			for(var/i = 1 to requirments[obj])
+				var/req_item = locate(obj) in get_turf(ritual_rune)
+				qdel(req_item)
 
 	if(ritual_flags & RITUAL_BLOODY)
 		if(ishuman(user))
 			var/mob/living/carbon/human/H = user
 			H.drip(5, get_turf(ritual_rune))
+			H.visible_message(SPAN_WARNING("[H] cuts his finger and lets some blood out!"))
+			sleep(5)
+			if(!check_cultists(ritual_rune))
+				return
 
 	if(ritual_flags & RITUAL_VERY_BLOODY)
 		if(ishuman(user))
 			var/mob/living/carbon/human/H = user
 			for(var/i = 1 to 3)
 				H.drip(5, get_turf(ritual_rune))
+			H.visible_message(SPAN_WARNING("[H] cuts his wrist and blood sprays from the cut!"))
+			sleep(5)
+			if(!check_cultists(ritual_rune))
+				return
+
+
+/mob/living/carbon/human/proc/zombify_cult()
+	src.rejuvenate()
+	ChangeToHusk()
+	mutations |= MUTATION_CLUMSY //No shooting zombies
+	src.visible_message("<span class='danger'>\The [src]'s skin decays before your very eyes!</span>", "<span class='cult italic'>Your entire body is ripe with pain as it is consumed down to flesh and bones. You feel that Nar'Sie, The Geometer of Blood called you from the world of dead to serve him!</span>")
+	if (src.mind)
+		GLOB.cult.add_antagonist(mind, ignore_role = 1, do_not_equip = 1)
+
+	log_admin("[key_name(src)] has transformed into a cult zombie!")
+	if (should_have_organ(BP_HEART))
+		vessel.add_reagent(/datum/reagent/blood, species.blood_volume - vessel.total_volume)
+	for (var/o in organs)
+		var/obj/item/organ/organ = o
+		organ.vital = 0
+		if(istype(organ, /obj/item/organ/internal))
+			if (!BP_IS_ROBOTIC(organ))
+				organ.rejuvenate(1)
+				organ.max_damage *= 3
+				organ.min_broken_damage = Floor(organ.max_damage * 0.75)
+		else
+			if (!BP_IS_ROBOTIC(organ))
+				organ.rejuvenate(1)
+				organ.max_damage *= 0.5
+				organ.min_broken_damage = Floor(organ.max_damage * 0.75)
+	playsound(get_turf(src), 'sound/hallucinations/wail.ogg', 20, 1)
+
+/mob
+	var/mirror_curse = 0 //If it isn't 0, victim will hear everybody as himself for a period of time(in eye_blink)
+
+	//ELAR YOU MOTHERFUCKER, FIND THE WAY TO MAKE EVERYBODY LOOK LIKE THAT ONE MOB
+
+/mob/living/carbon/human
+	var/list/bloodybond = list()
